refactor: 대기열 도메인 관련 메서드명 변경 등 전반적인 리팩토링
