+++ src/main/java/com/hhpl/concertreserve/domain/waitingqueue/WaitingQueueScheduler.java
@@ -10,24 +10,18 @@ import java.util.List;
 @RequiredArgsConstructor
 public class WaitingQueueScheduler {
 
-    private final WaitingQueueRepository waitingQueueRepository;
+    private final WaitingQueueService waitingQueueService;
+
 
     @Scheduled(fixedRate = 30000)
     public void activateWaitingQueuesForAllConcerts() {
-        List<Long> concertIds = waitingQueueRepository.findAllConcertIdsWithInactiveQueues();
-
-        concertIds.forEach(this::activateWaitingQueues);
+        waitingQueueService.activateInactiveQueuesForAllConcerts();
     }
 
-    public void activateWaitingQueues(Long concertId) {
-        int activationLimit = 5;
-
-        List<WaitingQueue> inactiveQueues = waitingQueueRepository.findInactiveQueuesForActivation(concertId,activationLimit);
-
-        inactiveQueues.forEach(queue -> {
-            queue.activate();
-            waitingQueueRepository.save(queue);
-        });
+    @Scheduled(fixedRate = 60000)
+    public void checkAndExpireWaitingQueue() {
+        waitingQueueService.expireActiveQueuesPastDeadline();
     }
-    
+
+
 }
